fix: fetchWpGlobalStyles cjs to esm
